Changed rotation and occlusion
Rotation now will move the image based on the direction of the hand tilt.
There was an issue where the thumb wasn't being picked up at 90 degree tilt.
A tilt of more than .05 in either direction will start rotating

Changed occlusion so that each pitch only changes one occlusion.
Previously if we had 3 images, we couldn't set an image to be inbetween.

Added a check that clears the pitch hand position when we stop using that position.